requirements.txt
streamlit
pandas
seaborn
numpy
matplotlib
scikit-learn
Pillow
